Modified Bubble.
